index.js contains graph.js instead of requiring it
